fix NewsControllerに統合
